requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
pydantic>=2.7
google-generativeai>=0.7.2
elevenlabs>=1.9.0
requests>=2.32
pymongo>=4.6
